feat!: Cleaner, lighter weight patch (#66)
## Summary

This is not the planned rewrite, but the codebase was in desperate need of an update. The patch build system with rollup was failing with newer versions and was hacky under the best of circumstances, anyway. This new version has a proper custom build system that produces a much leaner patch. Additionally, I reorganized the codebase, improved tests, and dropped old TS support. 

_Note: If you absolutely need it, it will still work with older TS. Simply fork and remove the version check_

## Changes

- Soft deprecated `--basedir` (use `--dir`)
- Zero bundled dependencies in patch (much lighter)
- ⚠️ Hard deprecated `--persist` option (use `package.json` -> `prepare` script)
- ⚠️ Requires TS >= 4.0

_(⚠️ denotes potential "breaking" change)_

## Development Changes

- Removed rollup and created light-weight custom build system
- Cleaned up file structure
- Improved test speed & methodology
- Changed patch detection signature